+++ test-module/src/test/java/com/ing/baker/JBakerTest.java
@@ -1,11 +1,12 @@
 package com.ing.baker;
 
 import com.google.common.collect.ImmutableList;
-import com.ing.baker.il.CompiledRecipe;
+import com.ing.baker.compiler.JavaCompiledRecipeTest;
 import com.ing.baker.compiler.RecipeCompiler;
-import com.ing.baker.runtime.core.BakerException;
+import com.ing.baker.il.CompiledRecipe;
 import com.ing.baker.recipe.javadsl.InteractionDescriptor;
 import com.ing.baker.recipe.javadsl.Recipe;
+import com.ing.baker.runtime.core.BakerException;
 import com.ing.baker.runtime.java_api.JBaker;
 import com.typesafe.config.ConfigFactory;
 import org.junit.Rule;
@@ -14,18 +15,18 @@ import org.junit.rules.ExpectedException;
 
 import java.util.UUID;
 
-import static com.ing.baker.JCompiledRecipeTest.setupComplexRecipe;
-import static com.ing.baker.JCompiledRecipeTest.setupSimpleRecipe;
+import static com.ing.baker.compiler.JavaCompiledRecipeTest.setupComplexRecipe;
+import static com.ing.baker.compiler.JavaCompiledRecipeTest.setupSimpleRecipe;
 import static org.junit.Assert.assertEquals;
 
 //TODO move to runtime, not possbile now because it references parts of the RecipeCompilerSpec
 public class JBakerTest {
 
     java.util.List<Object> implementationsList = ImmutableList.of(
-            new JCompiledRecipeTest.InteractionOneImpl(),
-            new JCompiledRecipeTest.InteractionTwo(),
-            new JCompiledRecipeTest.InteractionThreeImpl(),
-            new JCompiledRecipeTest.SieveImpl());
+            new JavaCompiledRecipeTest.InteractionOneImpl(),
+            new JavaCompiledRecipeTest.InteractionTwo(),
+            new JavaCompiledRecipeTest.InteractionThreeImpl(),
+            new JavaCompiledRecipeTest.SieveImpl());
 
     @Rule
     public final ExpectedException exception = ExpectedException.none();
@@ -36,7 +37,7 @@ public class JBakerTest {
         assertEquals(jBaker.getCompiledRecipe().getValidationErrors().size(), 0);
         UUID requestId = UUID.randomUUID();
         jBaker.bake(requestId);
-        jBaker.processEvent(requestId, new JCompiledRecipeTest.EventOne());
+        jBaker.processEvent(requestId, new JavaCompiledRecipeTest.EventOne());
         assertEquals("{RequestIDStringOne=" + requestId.toString() + "}", jBaker.getIngredients(requestId).toString());
     }
 
@@ -47,7 +48,7 @@ public class JBakerTest {
         assertEquals(jBaker.getCompiledRecipe().getValidationErrors().size(), 0);
         UUID requestId = UUID.randomUUID();
         jBaker.bake(requestId);
-        jBaker.processEvent(requestId, new JCompiledRecipeTest.EventOne());
+        jBaker.processEvent(requestId, new JavaCompiledRecipeTest.EventOne());
         assertEquals("{RequestIDStringOne=" + requestId.toString() + "}", jBaker.getIngredients(requestId).toString());
     }
 
@@ -63,13 +64,13 @@ public class JBakerTest {
         JBaker jBaker = new JBaker(RecipeCompiler.compileRecipe(setupComplexRecipe()), implementationsList);
         UUID processId = UUID.randomUUID();
         jBaker.bake(processId);
-        jBaker.processEvent(processId, new JCompiledRecipeTest.EventOne());
-        jBaker.processEvent(processId, new JCompiledRecipeTest.EventTwo());
+        jBaker.processEvent(processId, new JavaCompiledRecipeTest.EventOne());
+        jBaker.processEvent(processId, new JavaCompiledRecipeTest.EventTwo());
     }
 
     @Test
     public void shouldFailWhenSieveNotDefaultConstructor() throws BakerException {
-        Recipe recipe = setupComplexRecipe().withSieve(InteractionDescriptor.of(JCompiledRecipeTest.SieveImplWithoutDefaultConstruct.class));
+        Recipe recipe = setupComplexRecipe().withSieve(InteractionDescriptor.of(JavaCompiledRecipeTest.SieveImplWithoutDefaultConstruct.class));
         exception.expect(BakerException.class);
         CompiledRecipe compiledRecipe = RecipeCompiler.compileRecipe(recipe);
         JBaker jBaker = new JBaker(compiledRecipe, implementationsList);